Remove the confusing incident Reply page menu
The page menu was added along with refine page in 1e27969b79 so users
can easily jump between reply and refine pages. Since we switched to RT
search pages, the jump doesn't work any more because we didn't add
"Reply" on RT search page accordingly.

Currently we actually have 2 reply actions("Reply to Reporters" and
"Reply to All"), even though the menu remembers if it's "Reply to
Reporters" and "Reply to All" on current page, it still confuses people.